Update ntexplorehost.cpp
A little bit more coding. Added some good features like if the program is not in registry we add it and kill this instance of current program and run the exe in the desired location and we also start the log in that spot. Probably a little sloppy it is 2:26 am.